switching from domain module to domain package
