+++ app/src/main/java/com/kerollosragaie/mealzapp/di/RepositoryModule.kt
@@ -2,7 +2,7 @@ package com.kerollosragaie.mealzapp.di
 
 import com.kerollosragaie.data.remote.ApiServices
 import com.kerollosragaie.data.repository.MealsRepoImpl
-import com.kerollosragaie.domain.repository.MealsRepository
+import com.kerollosragaie.mealzapp.features.meals_menu.domain.repository.MealsRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
