+++ app/src/main/java/com/kerollosragaie/mealzapp/features/meals_menu/data/remote/ApiServices.kt
@@ -1,9 +1,9 @@
 package com.kerollosragaie.data.remote
 
-import com.kerollosragaie.domain.entity.CategoryResponse
+import com.kerollosragaie.mealzapp.features.meals_menu.domain.entity.CategoryResponse
 import retrofit2.http.GET
 
 interface ApiServices {
     @GET("categories.php")
-    suspend fun getMeals():CategoryResponse
+    suspend fun getMeals(): CategoryResponse
 }
