+++ app/src/main/java/com/kerollosragaie/mealzapp/features/meals_menu/data/repository/MealsRepoImpl.kt
@@ -1,8 +1,8 @@
 package com.kerollosragaie.data.repository
 
 import com.kerollosragaie.data.remote.ApiServices
-import com.kerollosragaie.domain.entity.CategoryResponse
-import com.kerollosragaie.domain.repository.MealsRepository
+import com.kerollosragaie.mealzapp.features.meals_menu.domain.entity.CategoryResponse
+import com.kerollosragaie.mealzapp.features.meals_menu.domain.repository.MealsRepository
 
 class MealsRepoImpl(private val apiServices: ApiServices) : MealsRepository {
     override suspend fun getMealsFromRemote(): CategoryResponse =
