+++ app/src/main/java/com/kerollosragaie/mealzapp/features/meals_menu/domain/entity/CategoryResponse.kt
@@ -1,0 +1,5 @@
+package com.kerollosragaie.mealzapp.features.meals_menu.domain.entity
+
+data class CategoryResponse(
+    val categories: List<Category>
+)
