+++ app/src/main/java/com/kerollosragaie/mealzapp/features/meals_menu/domain/repository/MealsRepository.kt
@@ -1,0 +1,7 @@
+package com.kerollosragaie.mealzapp.features.meals_menu.domain.repository
+
+import com.kerollosragaie.mealzapp.features.meals_menu.domain.entity.CategoryResponse
+
+interface MealsRepository {
+    suspend fun getMealsFromRemote(): CategoryResponse
+}
