+++ app/src/main/java/com/kerollosragaie/mealzapp/features/meals_menu/domain/usecases/GetMealsUseCaseImpl.kt
@@ -1,0 +1,7 @@
+package com.kerollosragaie.mealzapp.features.meals_menu.domain.usecases
+
+import com.kerollosragaie.mealzapp.features.meals_menu.domain.repository.MealsRepository
+
+class GetMealsUseCaseImpl(private val mealsRepository: MealsRepository) {
+  suspend operator fun invoke() = mealsRepository.getMealsFromRemote()
+}
