--- domain/src/main/java/com/kerollosragaie/domain/entity/CategoryResponse.kt
@@ -1,5 +1,0 @@
-package com.kerollosragaie.domain.entity
-
-data class CategoryResponse(
-    val categories: List<Category>
-)
