--- domain/src/main/java/com/kerollosragaie/domain/repository/MealsRepository.kt
@@ -1,7 +1,0 @@
-package com.kerollosragaie.domain.repository
-
-import com.kerollosragaie.domain.entity.CategoryResponse
-
-interface MealsRepository {
-    suspend fun getMealsFromRemote(): CategoryResponse
-}
