--- domain/src/main/java/com/kerollosragaie/domain/usecases/GetMealsUseCaseImpl.kt
@@ -1,7 +1,0 @@
-package com.kerollosragaie.domain.usecases
-
-import com.kerollosragaie.domain.repository.MealsRepository
-
-class GetMealsUseCaseImpl(private val mealsRepository: MealsRepository) {
-  suspend operator fun invoke() = mealsRepository.getMealsFromRemote()
-}
